chore: refactor pkgs
